GET '/auth/email-confirmation' initial setup + models + migrations.
