+++ api_server/auth/models/__init__.py
@@ -1,0 +1,5 @@
+from auth.models.email_confirmation_tokens import EmailConfirmationToken
+
+__all__ = [
+    'EmailConfirmationToken',
+]
